requirements.txt
-e .
funcy==1.8
influxdb==4.1.1
numpy==1.13.0
pandas==0.20.2
git+git://github.com/s4w3d0ff/python-poloniex.git@1e0cdcb757dee83cd28f85461e329d866a93a4a5#egg=poloniex
PyStaticConfiguration==0.10.3
requests==2.18.1
